graduate qwik
