+++ packages/create-cloudflare/src/templates.ts
@@ -19,7 +19,6 @@ import {
 	writeJSON,
 } from "helpers/files";
 import nextTemplateExperimental from "templates-experimental/next/c3";
-import qwikTemplateExperimental from "templates-experimental/qwik/c3";
 import solidTemplateExperimental from "templates-experimental/solid/c3";
 import analogTemplate from "templates/analog/c3";
 import angularTemplate from "templates/angular/c3";
@@ -173,7 +172,6 @@ export function getFrameworkMap({ experimental = false }): TemplateMap {
 	if (experimental) {
 		return {
 			next: nextTemplateExperimental,
-			qwik: qwikTemplateExperimental,
 			solid: solidTemplateExperimental,
 		};
 	} else {
