+++ packages/create-cloudflare/templates/qwik/pages/snippets/getPlatformProxy.ts
