+++ packages/create-cloudflare/templates/qwik/pages/templates/worker-configuration.d.ts
