+++ packages/create-cloudflare/templates/qwik/pages/templates/wrangler.jsonc
