+++ packages/create-cloudflare/templates/qwik/workers/snippets/getPlatformProxy.ts
