+++ packages/create-cloudflare/templates/qwik/workers/templates/public/.assetsignore
