+++ packages/create-cloudflare/templates/qwik/workers/templates/worker-configuration.d.ts
