+++ packages/create-cloudflare/templates/qwik/workers/templates/wrangler.jsonc
